requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
Flask-Migrate>=4.0
PyMySQL>=1.1
PyJWT>=2.8
pytest>=8.0
 SQLAlchemy>=2.0
 alembic>=1.13
 ultralytics
 --extra-index-url https://download.pytorch.org/whl/cpu
 torch==2.3.1+cpu
 torchvision==0.18.1+cpu
 opencv-python-headless>=4.10.0
dotenv